Refactored DANS bag validation, compiling again.
