+++ src/main/scala/nl.knaw.dans.easy.dd2d/DansDeposit2ToDataverseApp.scala
@@ -18,6 +18,7 @@ package nl.knaw.dans.easy.dd2d
 import java.io.PrintStream
 
 import better.files.File
+import nl.knaw.dans.easy.dd2d.dansbag.DansBagValidator
 import nl.knaw.dans.easy.dd2d.dataverse.DataverseInstance
 import nl.knaw.dans.easy.dd2d.queue.{ ActiveTaskQueue, PassiveTaskQueue }
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
@@ -28,18 +29,18 @@ import scala.util.Try
 class DansDeposit2ToDataverseApp(configuration: Configuration) extends DebugEnhancedLogging {
   private implicit val resultOutput: PrintStream = Console.out
   private val dataverse = new DataverseInstance(configuration.dataverse)
-  private val bagValidator = new BagValidator(configuration.validatorServiceUrl)
-  private val inboxWatcher = new InboxWatcher(new Inbox(configuration.inboxDir, dataverse))
+  private val dansBagValidator = new DansBagValidator(configuration.validatorServiceUrl)
+  private val inboxWatcher = new InboxWatcher(new Inbox(configuration.inboxDir, dansBagValidator, dataverse))
 
   def checkPreconditions(): Try[Unit] = {
     for {
-      _ <- bagValidator.checkConnection()
+      _ <- dansBagValidator.checkConnection()
       _ <- dataverse.checkConnection()
     } yield ()
   }
 
   def importDeposits(inbox: File): Try[Unit] = Try {
-    new InboxProcessor(new Inbox(inbox, dataverse)).process()
+    new InboxProcessor(new Inbox(inbox, dansBagValidator, dataverse)).process()
   }
 
   def start(): Try[Unit] = Try {
